requirements.txt
Appium-Python-Client==4.5.0
attrs==25.1.0
certifi==2024.12.14
cffi==1.17.1
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
pip==25.0
pycparser==2.22
PySocks==1.7.1
selenium==4.28.1
setuptools==68.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wheel==0.41.2
wsproto==1.2.0
